+++ src/components/ContentArea.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import Timeline from './Timeline/Timeline';
+import VideoEditor from './VideoEditor';
 
 function RouteContentArea() {
 	return (
 		<Switch>
-			<Route exact path="/editor" component={() => <Timeline />} />
+			<Route exact path="/editor" component={VideoEditor} />
 			<Route exact path="/chapters" component={() => <div>Capitulos</div>} />
 			<Route exact path="/documents" component={() => <div>Documentos</div>} />
 			<Route exact path="/branding" component={() => <div>Branding</div>} />
