+++ src/components/TimelineControl.module.scss
@@ -1,0 +1,107 @@
+.scrollbar {
+  display: flex;
+  /* flex-wrap: wrap; */
+  width: 86%;
+  height: 30px;
+  margin-left: 5%;
+  margin-top: 3px;
+
+  color: white;
+
+  /* overflow-y: hidden;
+          overflow-x: auto;
+          white-space: nowrap; */
+}
+
+.blackbox {
+  width: 10%;
+  min-width: 10%;
+  background-color: rgb(53, 52, 52);
+}
+
+.scrollbar__container {
+  display: flex;
+  flex-wrap: wrap;
+  width: 70%;
+}
+
+.timeline__scrollbar {
+  width: 100%;
+  min-width: 100%;
+  background-color: rgb(53, 52, 52);
+  color: white;
+  position: relative;
+
+  overflow-y: hidden;
+  overflow-x: auto;
+  white-space: nowrap;
+}
+
+.scrollbar__blackbox {
+  width: 100%;
+  min-width: 100%;
+  background-color: rgb(53, 52, 52);
+}
+
+.zoom__bar {
+  display: flex;
+  justify-content: space-around;
+
+  align-items: center;
+  background-color: rgb(53, 52, 52);
+  color: white;
+  width: 20%;
+}
+
+.btnContainer__icon {
+  height: 1.5rem;
+  width: 1.5rem;
+}
+
+.slider {
+  -webkit-appearance: none;
+  width: 100%;
+  height: 8px;
+  border-radius: 5px;
+  background: #d3d3d3;
+  outline: none;
+  opacity: 0.7;
+  -webkit-transition: 0.2s;
+  transition: opacity 0.2s;
+}
+
+.slider::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  appearance: none;
+  width: 14px;
+  height: 14px;
+  border-radius: 50%;
+  background: rgb(0, 174, 255);
+  cursor: pointer;
+}
+
+.slider::-moz-range-thumb {
+  width: 14px;
+  height: 14px;
+  border-radius: 50%;
+  background: rgb(0, 174, 255);
+  cursor: pointer;
+}
+
+.timeline__scrollbar::-webkit-scrollbar-track {
+  box-shadow: inset 0 0 6px rgba(80, 80, 80, 0.3);
+  -webkit-box-shadow: inset 0 0 6px rgba(80, 80, 80, 0.3);
+  /* border-radius: 10px; */
+  background-color: #666666;
+}
+
+.timeline__scrollbar::-webkit-scrollbar {
+  width: 1000px;
+  background-color: #ebebeb;
+}
+
+.timeline__scrollbar::-webkit-scrollbar-thumb {
+  box-shadow: inset 0 0 6px rgba(80, 80, 80, 0.3);
+  -webkit-box-shadow: inset 0 0 6px rgba(66, 62, 62, 0.3);
+  background-color: #d1cfcf;
+}
