+++ src/components/VideoEditor.module.scss
@@ -1,0 +1,23 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+*,
+*::before,
+*::after {
+  box-sizing: inherit;
+}
+
+html {
+  box-sizing: border-box;
+  /* font-size: 62.5%; 1rem = 10px, 10px/16px = 62.5% */
+}
+
+/*  GENERAL STYLE */
+
+.timeline__video-invisible {
+  /* display: flex;
+    flex-direction: column;
+    flex-wrap: wrap; */
+}
